requirements-ml-runtime.txt
# ML Runtime image dependencies
# This image runs inside Katib trial pods and Kubernetes training jobs.

# Object storage (MinIO / S3)
boto3>=1.34.0

# Data processing
pandas>=2.2.0

# ML training
scikit-learn>=1.5.0
joblib>=1.4.0

# Configuration
pyyaml>=6.0.0

# Kubernetes SDK (for reading Katib experiment results in train_model.py)
kubernetes>=28.0.0
